html/dom/elements/global-attributes/the-translate-attribute-010.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>translate-enabled state, inherited no</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/html5/dom.html#the-translate-attribute'>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<meta name="assert" content="If the translate attribute is set to no, javascript will detect the translation mode of elements inside that element with no translate flag as no-translate.">
<style type='text/css'>

		.test { border: 0; }
		.test div { color:white; padding: 5px; background-color: red; height: 90px; width: 90px; }

		</style>
</head>
<body>



<div class="test"><div id="box" translate="no">Fail <span id="spantest">&nbsp;</span></div></div>


<script>var pass = false; if (document.getElementById('spantest').translate === false) {  pass=true; }
if (pass) {  box.style.backgroundColor = '#00FF00';  box.style.color = 'white'; box.innerHTML = 'Pass'; box.style.height = '100px'; box.style.width = '100px'; }
test(function () {
    assert_true(pass);
	}, '');
</script><div id='log'></div></body>
</html>
